Splitting units into two files
